Update ArrayConfiguration to handle missing keys
It is possible that not all keys would be provided to the
ArrayConfiguration instance because the end developer wanted to use
browser default values. This was particularly true of the `max_age` key.
The ArrayConfiguration instance now handles when optional keys are not
provided and returns reasonable "null values".

If the ArrayConfiguration does not receive AT LEAST an array with keys
`origin` and `allowed_methods` an exception will be thrown because CORS
doesn't make sense without at least this information.

This addresses issue #6