Allow no methods to be defined for Origin (#21)
It was discovered through the course of documenting expected user cases
that a user might want to create a Configuration that specifies an
Origin is allowed to perform no HTTP methods. For example, if creating a
ConfigurationLoader and you encounter a Request you cannot reasonably
create a Configuration for you could create one on-the-fly for that
specific Origin that denies them access to any HTTP method.