Do not handle requests with no Origin header
Going through the ConfigurationLoader process and the necessary logic to
determine if CORS headers should be added for a Response if no Origin
header is present doesn't make sense. We know before any of the
processing that an Origin header with no value will never match any of
our expected logic and we can skip over this potentially costly step for
requests that do not require it.